--- 08. Advance Hooks & React 19/2. Use Context/2. With Use/components/Theme.jsx
@@ -1,28 +1,0 @@
-import { createContext, useState } from "react";
-import Card from "./Card";
-
-export const ThemeContext = createContext();
-
-const ThemeProvider = ({ children }) => {
-  const [theme, setTheme] = useState("light");
-
-  const toggleTheme = () => {
-    setTheme(theme === "light" ? "dark" : "light");
-  };
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
-
-const Theme = () => {
-  return (
-    <ThemeProvider>
-      <Card />
-    </ThemeProvider>
-  );
-};
-
-export default Theme;
